+++ backend/src/main/java/ssafy/a303/backend/property/dto/response/DetailResponseDto.java
@@ -12,8 +12,8 @@ public record DetailResponseDto(
 
         // 주소 좌표
         String address,
-        Integer latitude,
-        Integer longitude,
+        Double latitude,
+        Double longitude,
 
         //면적
         Double area,
