+++ backend/src/main/java/ssafy/a303/backend/property/dto/response/PropertyAddressResponseDto.java
@@ -4,7 +4,7 @@ public record PropertyAddressResponseDto(
         Integer propertySeq,
         String lessorNm,
         String address,
-        Integer latitude,
-        Integer longitude
+        Double latitude,
+        Double longitude
 ) {
 }
